Fix Html import error by restructuring layout
